+++ app/Http/Controllers/PatientController.php
@@ -7,7 +7,11 @@ use App\Patient;
 
 class PatientController extends Controller
 {
-    //
+	//
+	public function __construct()
+    {
+        $this->middleware('auth');
+    }
 	
 	public function create()
 	{
@@ -21,6 +25,17 @@ class PatientController extends Controller
 	public function store(Request $request)
 	{
 		$patient = new Patient;
+		$request->validate([
+			'patient_no' => ['required', 'unique:patients','regex:/^PID\-\d{5}$/'],
+			'name' => 'required|max:300',
+			'nric' => ['required','regex:/^\d{6}\-\d{2}\-\d{4}$/'],
+			'gender' => 'required',
+			'phone' => ['required','regex:/^\d{2,3}\-\d{6,10}$/'],
+			'address' => 'required|max:500',
+			'postcode' => 'required|numeric',
+			'city' => 'required',
+			'state' => 'required',
+		]);
 		$patient->fill($request->all());
 		$patient->save();
 	
@@ -48,12 +63,17 @@ class PatientController extends Controller
 	
 	public function edit($id)
 	{
-		$patient = Patient::find($id);
-		if(!$patient) throw new ModelNotFoundException;
+		if(\Gate::allows('isSuperAdmin')){
+			$patient = Patient::find($id);
+			if(!$patient) throw new ModelNotFoundException;
 
-		return view('patients.edit', [
-		'patient' => $patient
-		]);
+			return view('patients.edit', [
+			'patient' => $patient
+			]);
+		}else{
+			\Session::flash('msg', 'Unable to edit. ');
+			return \Redirect::back();
+		}
 	}
 	
 	public function update(Request $request, $id)
@@ -61,6 +81,17 @@ class PatientController extends Controller
 		$patient = Patient::find($id);
 		if(!$patient) throw new ModelNotFoundException;
 
+		$request->validate([
+			'patient_no' => ['required', 'regex:/^PID\-\d{5}$/'],
+			'name' => 'required|max:300',
+			'nric' => ['required','regex:/^\d{6}\-\d{2}\-\d{4}$/'],
+			'gender' => 'required',
+			'phone' => ['required','regex:/^\d{2,3}\-\d{6,10}$/'],
+			'address' => 'required|max:500',
+			'postcode' => 'required|numeric',
+			'city' => 'required',
+			'state' => 'required',
+		]);
 		$patient->fill($request->all());
 
 		$patient->save();
@@ -75,12 +106,17 @@ class PatientController extends Controller
      */	
 	public function destroy($id)
     {
-		$patient =Patient::where('id',$id)->first();
+		if(\Gate::allows('isSuperAdmin')){
+			$patient =Patient::where('id',$id)->first();
 
-		if ($patient != null) {
-			$patient->delete();
-			return redirect()->route('patient.index')->with('success','Patient deleted successfully');
+			if ($patient != null) {
+				$patient->delete();
+				return redirect()->route('patient.index')->with('success','Patient deleted successfully');
+			}
+			return redirect()->route('patient.index')->with(['message'=> 'Wrong ID!!']);
+		}else{
+			\Session::flash('msg', 'Unable to delete. ');
+			return \Redirect::back();
 		}
-		return redirect()->route('patient.index')->with(['message'=> 'Wrong ID!!']);
     }
 }
