+++ app/Http/Controllers/VisitController.php
@@ -8,6 +8,10 @@ use App\Visit;
 class VisitController extends Controller
 {
     //
+	public function __construct()
+    {
+        $this->middleware('auth');
+	}
 	
 	public function create()
 	{
@@ -21,6 +25,15 @@ class VisitController extends Controller
 	public function store(Request $request)
 	{
 		$visit = new Visit;
+		$request->validate([
+			'code' => ['required', 'unique:visits','regex:/^APT\-\d{5}$/'],
+			'time' => 'required',
+			'date' => 'required',
+			'doctor_id' => 'required',
+			'patient_id' => 'required',
+			'description' => 'required|max:500',
+			'statue' => 'required',
+		]);
 		$visit->fill($request->all());
 		$visit->save();
 	
@@ -39,6 +52,7 @@ class VisitController extends Controller
 	public function show($id)
 	{
 		$visit = Visit::find($id);
+		
 		if(!$visit) throw new ModelNotFoundException;
 	
 		return view('visits.show', [
@@ -48,12 +62,17 @@ class VisitController extends Controller
 	
 	public function edit($id)
 	{
-		$visit = Visit::find($id);
-		if(!$visit) throw new ModelNotFoundException;
+		if(\Gate::allows('isSuperAdmin')){
+			$visit = Visit::find($id);
+			if(!$visit) throw new ModelNotFoundException;
 
-		return view('visits.edit', [
-		'visit' => $visit
-		]);
+			return view('visits.edit', [
+			'visit' => $visit
+			]);
+		}else{
+			\Session::flash('msg', 'Unable to edit. ');
+			return \Redirect::back();
+		}
 	}
 	
 	public function update(Request $request, $id)
@@ -61,6 +80,15 @@ class VisitController extends Controller
 		$visit = Visit::find($id);
 		if(!$visit) throw new ModelNotFoundException;
 
+		$request->validate([
+			'code' => ['required','regex:/^APT\-\d{5}$/'],
+			'time' => 'required',
+			'date' => 'required',
+			'doctor_id' => 'required',
+			'patient_id' => 'required',
+			'description' => 'required|max:500',
+			'statue' => 'required',
+		]);
 		$visit->fill($request->all());
 
 		$visit->save();
@@ -75,13 +103,18 @@ class VisitController extends Controller
      */	
 	public function destroy($id)
     {
-		$visit =Visit::where('id',$id)->first();
+		if(\Gate::allows('isSuperAdmin')){
+			$visit =Visit::where('id',$id)->first();
 
-		if ($visit != null) {
-			$visit->delete();
-			return redirect()->route('visit.index')->with('success','Visit deleted successfully');
+			if ($visit != null) {
+				$visit->delete();
+				return redirect()->route('visit.index')->with('success','Visit deleted successfully');
+			}
+			return redirect()->route('visit.index')->with(['message'=> 'Wrong ID!!']);
+		}else{
+			\Session::flash('msg', 'Unable to delete. ');
+			return \Redirect::back();
 		}
-		return redirect()->route('visit.index')->with(['message'=> 'Wrong ID!!']);
     }
 
 }
